+++ projects/egeo/src/lib/st-dynamic-table/spec/st-dynamic-table.component.spec.ts
@@ -397,6 +397,28 @@ describe('StDynamicTableComponent', () => {
       });
    });
 
+   describe('it should fix the hover menu of the current focused row in order to be displayed although table width is larger than screen´s one', () => {
+      it('if input "stickyHoverMenu" is not specified, hover menu is not fixed', () => {
+         fixture.detectChanges();
+
+         expect(fixture.nativeElement.querySelector('st-table').classList).not.toContain('sticky-hover-menu');
+      });
+
+      it('if input "stickyHoverMenu" is true, hover menu is fixed', () => {
+         component.stickyHoverMenu = true;
+         fixture.detectChanges();
+
+         expect(fixture.nativeElement.querySelector('st-table').classList).toContain('sticky-hover-menu');
+      });
+
+      it('if input "stickyHoverMenu" is false, hover menu is not fixed', () => {
+         component.stickyHoverMenu = false;
+         fixture.detectChanges();
+
+         expect(fixture.nativeElement.querySelector('st-table').classList).not.toContain('sticky-hover-menu');
+      });
+   });
+
    it('Custom classes can be added to the table', () => {
       fixture.detectChanges();